Create Median
A median is a numerical value separating the upper half of a sorted list of numbers from the lower half. In list where there are an odd number of entities, the median is the number found in the middle of the list. If the list contains an even number of entities, then there is no single middle value, instead the median becomes the average of the two numbers found in the middle. For this mission, you are given a non-empty list of integers (int). With it, you must separate the upper half of the numbers from the lower half and find the median.